requirements.txt
click>=8.1.7
jinja2>=3.1.2
pyyaml>=6.0.1
requests>=2.31.0
colorama>=0.4.6
rich>=13.7.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
typing-extensions>=4.8.0
toml>=0.10.2
libcst>=1.0.0
astor>=0.8.1
